Minor style improvements:
* Remove misleading type annotations (most of it in fact)
* Move `PycudaGpu` into user space
* Ensure that `devices_taken` and `placeholders` are always present.
  This allows deletion of `GPUOwner` even when it failed to allocate the
  GPUs.
